requirements.txt
# requirements.txt
requests
psycopg2-binary